Singlemovie page (#19)
* product page updates

* added product page

grabs whatever is written after /browse in address and checks if there is a movie of that title in the testData

* cleanup of product page

cleanup and added youtube link to test trailers

* added product page

grabs whatever is written after /browse in address and checks if there is a movie of that title in the testData

* cleanup of product page

cleanup and added youtube link to test trailers

* product page updates

* updated dummy data

added fields needed for product detail page and changes in pages to match prisma model

* Moviedetail page updated.

Moviedetail done with partial dummy data, need to be updated with proper data later.
Genre is fetched and incorporated as of this commit

* params fix

removed an unnecessary toString in page
Also found an error in moviedetailpage where "id" should been "title". That entire function will need rework with real data though